pyproject.toml
[project]
name = "copilot-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Kunal7069",email = "jaink7069@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "websockets (>=15.0.1,<16.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
